+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.3.14](https://github.com/duality-labs/duality-web-app/compare/v0.3.13...v0.3.14) (2023-09-24)
+
+
+### Features
+
+* improve IBC bridging toast notifications ([#436](https://github.com/duality-labs/duality-web-app/issues/436)) ([df65f94](https://github.com/duality-labs/duality-web-app/commit/df65f94ef73a513cc12d0c6282802770eda89084))
+
+## [0.3.13](https://github.com/duality-labs/duality-web-app/compare/v0.3.12...v0.3.13) (2023-09-24)
+
+
+### Features
+
+* abstract consistent transaction notification toasts ([#434](https://github.com/duality-labs/duality-web-app/issues/434)) ([0de982d](https://github.com/duality-labs/duality-web-app/commit/0de982d6676edc32e0a3a4769799a5c9928e0942))
+
 ## [0.3.12](https://github.com/duality-labs/duality-web-app/compare/v0.3.11...v0.3.12) (2023-09-21)